change pages position
